refactor: change package structure to domain-based
처음에는 MVC 패턴을 위해 패키지 구조를 Model, View, Controller로 만들고
개발을 시작했습니다, 하지만 개발 도중 패키지 구조와 MVC 패턴은 아무런
관계가 없다는 것을 깨닫게 되었습니다. 기존의 패키지 구조로 계속 개발을 이어갈 경우
Model에는 RacingGame과 Car가 겹치게 되는데 이 경우 Car의 strategy를 Model 패키지에
두기엔 RacingGame과는 관련이 없어서 무언가 일관성이 떨어지고 도메인별 응집도가
낮아진다고 생각했습니다. 따라서 도메인별 응집도를 높이기 위해 계층형 구조에서
도메인형 구조로 패키지 구조를 변경했습니다.
