+++ src/main/java/racingcar/car/CarFactory.java
@@ -1,7 +1,8 @@
-package racingcar.model;
+package racingcar.car;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.car.strategy.MoveStrategy;
 
 public class CarFactory {
     public List<Car> createCars(List<String> carNames, MoveStrategy moveStrategy) {
