+++ src/main/java/racingcar/game/RacingGameController.java
@@ -1,0 +1,4 @@
+package racingcar.game;
+
+public class RacingGameController {
+}
